reorganizing go files
