--- go/src/TODO.md
@@ -1,11 +1,0 @@
-# TODO
-
-[x] ignore .gitignore paths
-[x] parse git blame
-[x] pretty output
-[x] shorten filepath
-[x] plain style support
-[x] use plain style when output is redirected
-[ ] CLI options
-[ ] test cross-platform compatibility
-[ ] config wizard
