--- pyproject.toml
@@ -1,24 +1,0 @@
-[build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
-
-[project.scripts]
-listme = "listme.parser:main"
-listme-config = "listme.config:wizard"
-
-[project]
-name = "listme"
-version = "0.2"
-authors = [
-  { name="Matheus Pedroni", email="pnmatheus@protonmail.com" },
-]
-description = "Summarize you FIXME, TODO, XXX (and other tags) comments so you don't forget them."
-readme = "README.md"
-license = { file="LICENSE" }
-dependencies = ["rich>=12.6.0", "platformdirs>=2.6.0"]
-requires-python = ">=3.8"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
-    "Operating System :: Unix",
-]
