--- src/listme/_main.py
@@ -1,10 +1,0 @@
-"""
-Minimal script to run listme without installing it.
-
-Used for testing purposes.
-"""
-
-from listme.parser import main
-
-if __name__ == "__main__":
-    main()
